dom/tests/mochitest/pointerlock/file_mouselocklost.html

<!DOCTYPE HTML>
<html>
  <!-- https://bugzilla.mozilla.org/show_bug.cgi?id=633602 -->
  <head>
    <title>
      Test for Bug 633602
    </title>
    <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js">
    </script>
    <script type="text/javascript" src="/tests/SimpleTest/EventUtils.js">
    </script>
    <script type="application/javascript" src="mouselock_util.js">
    </script>
    <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  </head>
  <body>
    <a target="_blank" 
      href="https://bugzilla.mozilla.org/show_bug.cgi?id=633602">
      Mozilla Bug 633602
    </a>
    <div id="div">
    </div>
    <pre id="test">
      <script type="application/javascript">
        /** Test for Bug 633602 **/
        /** This test is to verify that the mouselocklost event fired by
            the browser has the bubbles attribute set to true, and
            the target attribute is the correct element **/
        SimpleTest.waitForExplicitFinish();

        var pointer = navigator.mozPointer,
            div = document.getElementById("div"),
            firstTest = true;

        div.addEventListener("mozpointerlocklost", function(e) {
          if (firstTest) {
            /*  First test scenario: calling pointer.unlock() and testing the
              mouselockevent attributes */
            firstTest = false;
            ok(e.bubbles, "Manual unlock request - The mozpointerlocklost event " +
              "should have it's bubbles attribute set to true");
            is(e.target.id, "div", "target should be div");
            // Pending no problems, begin second test
            pointer.lock(div, function() {
              document.mozCancelFullScreen();
            });
          }          
          else if (!firstTest) {
            /*  Second test scenario: calling document.mozCancelFullScreen() 
                and testing the mouselockevent attributes */
            ok(e.bubbles, "mozCancelFullScreen request - The mouselocklost " +
              "event should have it's bubbles attribute set to true");
            is(e.target.id, "div", "target should be div");
            SimpleTest.finish();
          }
        }, false);

        document.addEventListener("mozfullscreenchange", function() {
          if (document.mozFullScreen) {
            // If we've locked on to the div element, proceed with test
            if (document.mozFullScreenElement === div) {
              pointer.lock(div, function(){
                pointer.unlock();
              });
            }
          }
        }, false);

        div.mozRequestFullScreen();
      </script>
    </pre>
  </body>
</html>
